requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pymongo>=4.5.0
pandas>=2.0.0
streamlit>=1.10.0
jupyter>=1.0.0